+++ Skvortsov.Vladimir/lab2/src/client_queue.cpp
@@ -6,46 +6,6 @@
 #include "conn_queue.hpp"
 #include "ChatWindow.hpp"
 
-ChatWindow::ChatWindow(QWidget* parent) : QMainWindow(parent), host_conn(nullptr), client_conn(nullptr) {
-  setup_ui();
-  setup_conn();
-  setup_timers();
-};
-
-ChatWindow::~ChatWindow() {
-  if (host_conn) {
-    delete host_conn;
-  }
-  if (client_conn) {
-    delete client_conn;
-  }
-};
-
-void ChatWindow::setup_ui() {
-  QWidget* central_widget = new QWidget(this);
-  setCentralWidget(central_widget);
-
-  chat_display = new QTextEdit(this);
-  chat_display->setReadOnly(true);
-
-  msg_input = new QLineEdit(this);
-  send_btn = new QPushButton("Send", this);
-
-  QVBoxLayout* mainLayout = new QVBoxLayout();
-  QHBoxLayout* inputLayout = new QHBoxLayout();
-
-  inputLayout->addWidget(msg_input);
-  inputLayout->addWidget(send_btn);
-
-  mainLayout->addWidget(chat_display);
-  mainLayout->addLayout(inputLayout);
-
-  central_widget->setLayout(mainLayout);
-
-  connect(send_btn, &QPushButton::clicked, this, &ChatWindow::send_msg);
-  connect(msg_input, &QLineEdit::returnPressed, this, &ChatWindow::send_msg);
-};
-
 void ChatWindow::setup_conn() {
   std::ifstream pid_file("host_pid.txt");
   if (!pid_file) {
@@ -76,12 +36,6 @@ void ChatWindow::setup_conn() {
     QMessageBox::critical(this, "Error", "Failed to open FIFO for reading");
     return;
   }
-};
-
-void ChatWindow::setup_timers() {
-  QTimer* timer = new QTimer(this);
-  connect(timer, &QTimer::timeout, this, &ChatWindow::read_msg);
-  timer->start(10);
 };
 
 void ChatWindow::send_msg() {
